migrating Preferences class into stand-alone version
